chore: tooling and scafolding
