+++ .github/CODEOWNERS
@@ -1,1 +1,1 @@
-*       @sl-julienamblard
+*       @julien-amblard
