+++ .prettierrc
@@ -1,18 +1,11 @@
 {
-	"arrowParens": "avoid",
-	"bracketSpacing": true,
-	"htmlWhitespaceSensitivity": "css",
-	"insertPragma": false,
-	"jsxBracketSameLine": false,
-	"jsxSingleQuote": false,
-	"printWidth": 120,
-	"proseWrap": "preserve",
-	"quoteProps": "as-needed",
-	"requirePragma": false,
-	"semi": false,
 	"singleQuote": false,
-	"tabWidth": 4,
-	"trailingComma": "es5",
-	"useTabs": true,
-	"vueIndentScriptAndStyle": false
+	"semi": false,
+	"printWidth": 80,
+	"endOfLine": "auto",
+	"tabWidth": 2,
+	"quoteProps": "as-needed",
+	"trailingComma": "none",
+	"bracketSpacing": true,
+	"arrowParens": "always"
 }
