+++ __tests__/sort.test.ts
@@ -1,0 +1,14 @@
+import { sortByCount } from "../src/helpers/sort"
+
+describe("sort", () => {
+  test("should return negative value", () => {
+    expect(sortByCount({ misspelledCount: 1 }, { misspelledCount: 2 })).toEqual(
+      -1
+    )
+  })
+  test("should return positive value", () => {
+    expect(sortByCount({ misspelledCount: 2 }, { misspelledCount: 1 })).toEqual(
+      1
+    )
+  })
+})
