+++ package.json
@@ -2,33 +2,30 @@
 	"name": "email-misspelled",
 	"description": "Check misspell email's domain and return a list of matching domain suggestions",
 	"main": "lib/index.js",
+	"types": "lib/index.d.ts",
 	"scripts": {
-		"dev": "webpack-dev-server --mode development --config webpack.dev.ts",
-		"test": "yarn test:watch",
-		"test:ci": "yarn jest --config=./tests/jest.config.js",
-		"test:watch": "yarn jest --config=./tests/jest.config.js --verbose --watchAll",
-		"test:coverage": "yarn jest --config=./tests/jest.config.js --verbose --coverage",
+		"test": "vitest",
+		"test:ci": "vitest run",
+		"test:coverage": "jest --config=./tests/jest.config.js --verbose --coverage",
 		"test:coverage:report": "codecov --token=$CODECOV_TOKEN",
 		"clear": "rimraf ./lib && rimraf ./domains && rimraf ./typings",
-		"build:all": "yarn clear && yarn build:webpack && yarn build:transpile && yarn build:domains && yarn build:types && yarn build:domains:copy && yarn build:types:copy",
-		"build:transpile": "babel src/index.ts --out-dir lib --extensions .ts",
-		"build:domains": "babel src/domains --out-dir lib/domains --extensions .ts",
-		"build:domains:copy": "cpy lib/domains domains",
-		"build:webpack": "webpack --config webpack.config.ts",
-		"build:types": "tsc",
-		"build:types:copy": "cpy lib/typings typings",
-		"prepublishOnly": "yarn test:ci && yarn build:all"
+		"build": "npm run clear && npm run transpile",
+		"transpile": "tsc --project ./tsconfig.json",
+		"prepublishOnly": "npm run test:ci && npm run build",
+		"prepare": "husky",
+		"lint:check": "eslint . --ignore-path .gitignore",
+		"lint:fix": "eslint . --ignore-path .gitignore --fix",
+		"format:check": "prettier --config .prettierrc \"src/**/*.ts\" --check",
+		"format:fix": "prettier --config .prettierrc \"src/**/*.ts\" --write"
+	},
+	"lint-staged": {
+		"src/**/*.{ts,tsx}": [
+			"npm run lint:fix",
+			"npm run format:fix"
+		]
 	},
 	"files": [
-		"/lib/index.js",
-		"/lib/index.d.ts",
-		"/lib/emailMisspelled.js",
-		"/lib/emailMisspelled.d.ts",
-		"/lib/emailMisspelled.web.js",
-		"/lib/domains",
-		"/lib/typings",
-		"/domains",
-		"/typings"
+		"/lib"
 	],
 	"repository": {
 		"type": "git",
@@ -57,30 +54,23 @@
 		"form"
 	],
 	"devDependencies": {
-		"@babel/cli": "^7.7.4",
-		"@babel/core": "^7.7.4",
-		"@babel/plugin-proposal-class-properties": "^7.7.4",
-		"@babel/plugin-proposal-object-rest-spread": "^7.7.4",
-		"@babel/plugin-proposal-optional-chaining": "^7.8.3",
-		"@babel/polyfill": "^7.7.0",
-		"@babel/preset-env": "^7.7.4",
-		"@babel/preset-typescript": "^7.8.3",
-		"@types/jest": "^26.0.0",
-		"@types/webpack": "^4.41.7",
-		"babel-jest": "^26.1.0",
-		"babel-loader": "^8.0.6",
-		"codecov": "^3.7.2",
-		"cpy-cli": "^3.1.1",
-		"css-loader": "^3.2.0",
-		"html-webpack-plugin": "^4.3.0",
-		"jest": "^26.1.0",
-		"node-sass": "^7.0.0",
-		"sass-loader": "^9.0.2",
-		"style-loader": "^1.0.1",
-		"ts-jest": "^26.1.1",
-		"typescript": "^4.2.0",
-		"webpack": "^4.43.0",
-		"webpack-cli": "^3.3.10",
-		"webpack-dev-server": "^3.11.2"
+		"@commitlint/cli": "^19.0.3",
+		"@commitlint/config-conventional": "^19.0.3",
+		"@typescript-eslint/eslint-plugin": "^7.1.0",
+		"@typescript-eslint/parser": "^7.1.0",
+		"eslint": "^8.57.0",
+		"eslint-config-airbnb": "^19.0.4",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-config-standard": "^17.1.0",
+		"eslint-import-resolver-typescript": "^3.6.1",
+		"eslint-plugin-import": "^2.29.1",
+		"eslint-plugin-unicorn": "^51.0.1",
+		"husky": "^9.0.11",
+		"lint-staged": "^15.2.2",
+		"prettier": "^3.2.5",
+		"rimraf": "^5.0.5",
+		"typescript": "^5.3.3",
+		"vite": "^5.1.4",
+		"vitest": "^1.3.1"
 	}
 }
