+++ src/domains/all.ts
@@ -1,11 +1,20 @@
-import aol from "./aol"
-import hotmail from "./hotmail"
-import live from "./live"
-import outlook from "./outlook"
-import more from "./more"
-import top100 from "./top100"
-import yahoo from "./yahoo"
-import { DomainList } from "../typings"
+import type { DomainList } from "../typings"
+import { aol } from "./aol"
+import { hotmail } from "./hotmail"
+import { live } from "./live"
+import { outlook } from "./outlook"
+import { more } from "./more"
+import { top100 } from "./top100"
+import { yahoo } from "./yahoo"
 
-export const all: DomainList = [...new Set([...top100, ...aol, ...hotmail, ...live, ...outlook, ...more, ...yahoo])]
-export default all
+export const all: DomainList = [
+  ...new Set([
+    ...top100,
+    ...aol,
+    ...hotmail,
+    ...live,
+    ...outlook,
+    ...more,
+    ...yahoo
+  ])
+]
