+++ src/domains/index.ts
@@ -1,9 +1,9 @@
-export { aol } from "./aol"
-export { hotmail } from "./hotmail"
-export { live } from "./live"
-export { outlook } from "./outlook"
-export { microsoft } from "./microsoft"
-export { more } from "./more"
-export { top100 } from "./top100"
-export { yahoo } from "./yahoo"
-export { all } from "./all"
+export * from "./aol"
+export * from "./hotmail"
+export * from "./live"
+export * from "./outlook"
+export * from "./microsoft"
+export * from "./more"
+export * from "./top100"
+export * from "./yahoo"
+export * from "./all"
