+++ src/domains/microsoft.ts
@@ -1,6 +1,6 @@
-import hotmail from "./hotmail"
-import live from "./live"
-import outlook from "./outlook"
-import { DomainList } from "../typings"
+import type { DomainList } from "../typings"
+import { hotmail } from "./hotmail"
+import { live } from "./live"
+import { outlook } from "./outlook"
 
 export const microsoft: DomainList = [...hotmail, ...live, ...outlook]
