+++ src/helpers/contains-one-at.ts
@@ -1,0 +1,2 @@
+export const containsOneAt = (email: string = ""): boolean =>
+  email.replace(/[^@]/g, "").length === 1
