--- src/helpers/containsOneAt.ts
@@ -1,1 +1,0 @@
-export const containsOneAt = (email: string = ""): boolean => email.replace(/[^@]/g, "").length === 1
