+++ src/helpers/corrector.ts
@@ -1,13 +1,16 @@
-import { Result } from "../typings"
+import type { Result } from "../typings"
+
 export interface Corrector {
-	(value: Result): Result
+  (value: Result): Result
 }
 interface CorrectorConfig {
-	(email: string): Corrector
+  (email: string): Corrector
 }
-export const corrector: CorrectorConfig = email => value => {
-	if (!!!email || !!!value.suggest) return value
-	value.corrected = email.replace(/@.*$/, `@${value.suggest}`)
-	value.original = email
-	return value
+export const corrector: CorrectorConfig = (email) => (value) => {
+  if (!email || !value.suggest) return value
+  return {
+    ...value,
+    corrected: email.replace(/@.*$/, `@${value.suggest}`),
+    original: email
+  }
 }
