+++ src/helpers/domain-mapper.ts
@@ -1,0 +1,11 @@
+import type { Result } from "../typings"
+
+export interface DomainMapper {
+  (suggest: string): Result
+}
+export const domainMapper: DomainMapper = (suggest) => ({
+  suggest,
+  corrected: "",
+  original: "",
+  misspelledCount: 0
+})
