+++ src/helpers/get-domain.ts
@@ -1,0 +1,2 @@
+export const getDomain = (email: string = ""): string =>
+  email.includes("@") ? email.replace(/.*@/, "") : ""
