--- src/helpers/getDomain.ts
@@ -1,1 +1,0 @@
-export const getDomain = (email: string = ""): string => (email.includes("@") ? email.replace(/.*@/, "") : "")
