+++ src/helpers/sort.ts
@@ -1,4 +1,4 @@
 interface Sort {
-	(a: { misspelledCount: number }, b: { misspelledCount: number }): number
+  (a: { misspelledCount: number }, b: { misspelledCount: number }): number
 }
 export const sortByCount: Sort = (a, b) => a.misspelledCount - b.misspelledCount
