+++ src/index.ts
@@ -1,4 +1,14 @@
+/* eslint import/no-default-export: 0 */
+import { emailMisspelled } from "./email-misspelled"
+
 export * from "./domains"
-import { emailMisspelled } from "./emailMisspelled"
-export type { DomainList, Result, EmailMisspelled, EmailMisspelledConfig, EmailMisspelledConstructor } from "./typings"
+
+export type {
+  DomainList,
+  Result,
+  EmailMisspelled,
+  EmailMisspelledConfig,
+  EmailMisspelledConstructor
+} from "./typings"
+
 export default emailMisspelled
