--- tests/src/containsOneAt.test.ts
@@ -1,14 +1,0 @@
-import { containsOneAt } from "../../src/helpers/containsOneAt"
-
-describe("containsOneAt : ", () => {
-	test("should return true", () => {
-		expect(containsOneAt("zefzef@zefzef.com")).toBe(true)
-		expect(containsOneAt("@")).toBe(true)
-	})
-	test("should return false", () => {
-		expect(containsOneAt("zefzef@zefz@ef.com")).toBe(false)
-		expect(containsOneAt("zefzefef.com")).toBe(false)
-		expect(containsOneAt("")).toBe(false)
-		expect(containsOneAt()).toBe(false)
-	})
-})
