--- tests/src/sort.test.ts
@@ -1,10 +1,0 @@
-import { sortByCount } from "../../src/helpers/sort"
-
-describe("sort", () => {
-	test("should return negative value", () => {
-		expect(sortByCount({ misspelledCount: 1 }, { misspelledCount: 2 })).toEqual(-1)
-	})
-	test("should return positive value", () => {
-		expect(sortByCount({ misspelledCount: 2 }, { misspelledCount: 1 })).toEqual(1)
-	})
-})
