+++ vitest.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from "vitest/config"
+
+// eslint-disable-next-line
+export default defineConfig({
+  test: {
+    globals: true,
+    environment: "node"
+  }
+})
